requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
